refactored two decimal

+++ berkel/UI/Modules/Jobi/Order/OrderViewController/SubControllers/OrderPaymentViewController/Models/OrderPaymentModel.swift
@@ -17,7 +17,7 @@ struct OrderPaymentModel: Codable {
     let customerName: String
 
     let date: String?
-    let payment: Int
+    let payment: Double
     var faturaNo: String?
     let description: String?
 }
